pyproject.toml
[project]
name = "geosignal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.26.0",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "mcp>=1.26.0",
    "openai>=2.24.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]
